Removed user setting for file_access_delay. App will now wait 10ms before accessing file. On access error app will retry 9 additonal times with a delay of (10ms * retry number) - so 10ms, 20ms, 30ms, etc. on each retry up to 90ms.
Added milliseconds to log time when 'log everything' is checked.

Added button to open log file from settings page.

Changed behavior so the lbl_errors.txt is cleared when user clicks on the message to view the log file. Error counter is NOT reset. This way once the user has clicked through to look at the log they do not have to remember the error count to know if a new error occurred. If lbl_errors has text then there has been an error since the last time the user clicked on the label to open the log.

Changed version on Overview tab to 1.65.

Changed published version in Visual Studio to 1.65.0.2